fix: move getDocumentTitle to rout util folder
