+++ packages/discovery-react-components/src/components/SearchResults/utils/index.ts
@@ -1,4 +1,3 @@
 export { findCollectionName } from './findCollectionName';
 export { getDisplaySettings } from './getDisplaySettings';
 export { findTablesWithoutResults } from './findTablesWithoutResults';
-export { getDocumentTitle } from './getDocumentTitle';
