+++ packages/discovery-react-components/src/utils/__tests__/getDocumentTitle.test.tsx
@@ -1,4 +1,4 @@
-import { getDocumentTitle } from '../index';
+import { getDocumentTitle } from '../getDocumentTitle';
 
 describe('getDocumentTitle', () => {
   test('returns result title', () => {
