+++ packages/discovery-react-components/src/utils/getDocumentTitle.ts
